requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-rest-multiple-models==2.1.3
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.1
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.5
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
